server/ai_services/requirements.txt
# Core AI Libraries
openai>=0.27.0
torch>=2.0.0
transformers>=4.30.0
whisper-openai>=20230314
elevenlabs>=0.2.24

# OCR and Image Processing
opencv-python>=4.7.0
pytesseract>=0.3.10
Pillow>=9.5.0
numpy>=1.24.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1

# Machine Learning Utilities
scikit-learn>=1.2.0
lime>=0.2.0
shap>=0.41.0

# NLP and Translation
sentencepiece>=0.1.99
sacrebleu>=2.3.1

# Data Management
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Type Checking
mypy>=1.3.0
types-requests>=2.31.0
types-setuptools>=68.0.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Documentation
Sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Development Tools
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
pre-commit>=3.3.3